Refactored provider classes from client.ts
