+++ Extension/src/LanguageServer/client.ts
@@ -6,7 +6,8 @@
 
 import * as path from 'path';
 import * as vscode from 'vscode';
-import * as editorConfig from 'editorconfig';
+import {FoldingRangeProvider, SemanticTokensProvider, DocumentFormattingEditProvider, DocumentRangeFormattingEditProvider, OnTypeFormattingEditProvider} from './providers';
+
 import {
     LanguageClient, LanguageClientOptions, ServerOptions, NotificationType, TextDocumentIdentifier,
     RequestType, ErrorAction, CloseAction, DidOpenTextDocumentParams, Range, Position, DocumentFilter
@@ -56,8 +57,8 @@ let debugChannel: vscode.OutputChannel;
 let diagnosticsCollection: vscode.DiagnosticCollection;
 let workspaceDisposables: vscode.Disposable[] = [];
 let workspaceReferences: refs.ReferencesManager;
-const openFileVersions: Map<string, number> = new Map<string, number>();
-const cachedEditorConfigSettings: Map<string, any> = new Map<string, any>();
+export const openFileVersions: Map<string, number> = new Map<string, number>();
+export const cachedEditorConfigSettings: Map<string, any> = new Map<string, any>();
 
 export function disposeWorkspaceData(): void {
     workspaceDisposables.forEach((d) => d.dispose());
@@ -299,7 +300,7 @@ interface DidChangeConfigurationParams extends WorkspaceFolderParams {
     settings: any;
 }
 
-interface FormatParams {
+export interface FormatParams {
     uri: string;
     range: Range;
     character: string;
@@ -313,12 +314,12 @@ interface TextEdit {
     newText: string;
 }
 
-interface GetFoldingRangesParams {
+export interface GetFoldingRangesParams {
     uri: string;
     id: number;
 }
 
-enum FoldingRangeKind {
+export enum FoldingRangeKind {
     None = 0,
     Comment = 1,
     Imports = 2,
@@ -339,7 +340,7 @@ interface AbortRequestParams {
     id: number;
 }
 
-interface GetSemanticTokensParams {
+export interface GetSemanticTokensParams {
     uri: string;
     id: number;
 }
@@ -403,9 +404,9 @@ const GetDiagnosticsRequest: RequestType<void, GetDiagnosticsResult, void, void>
 const GetCodeActionsRequest: RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void> = new RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void>('cpptools/getCodeActions');
 const GetDocumentSymbolRequest: RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void> = new RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void>('cpptools/getDocumentSymbols');
 const GetSymbolInfoRequest: RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void> = new RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void>('cpptools/getWorkspaceSymbols');
-const GetFoldingRangesRequest: RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void> = new RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void>('cpptools/getFoldingRanges');
-const GetSemanticTokensRequest: RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void> = new RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void>('cpptools/getSemanticTokens');
-const DocumentFormatRequest: RequestType<FormatParams, TextEdit[], void, void> = new RequestType<FormatParams, TextEdit[], void, void>('cpptools/format');
+export const GetFoldingRangesRequest: RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void> = new RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void>('cpptools/getFoldingRanges');
+export const GetSemanticTokensRequest: RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void> = new RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void>('cpptools/getSemanticTokens');
+export const DocumentFormatRequest: RequestType<FormatParams, TextEdit[], void, void> = new RequestType<FormatParams, TextEdit[], void, void>('cpptools/format');
 // Notifications to the server
 const DidOpenNotification: NotificationType<DidOpenTextDocumentParams, void> = new NotificationType<DidOpenTextDocumentParams, void>('textDocument/didOpen');
 const FileCreatedNotification: NotificationType<FileChangedParams, void> = new NotificationType<FileChangedParams, void>('cpptools/fileCreated');
@@ -464,7 +465,6 @@ interface ReferencesCancellationState {
 
 const referencesPendingCancellations: ReferencesCancellationState[] = [];
 
-let abortRequestId: number = 0;
 
 class ClientModel {
     public isTagParsing: DataBinding<boolean>;
@@ -569,287 +569,6 @@ export function createNullClient(): Client {
     return new NullClient();
 }
 
-class FoldingRangeProvider implements vscode.FoldingRangeProvider {
-    private client: DefaultClient;
-    constructor(client: DefaultClient) {
-        this.client = client;
-    }
-    provideFoldingRanges(document: vscode.TextDocument, context: vscode.FoldingContext,
-        token: vscode.CancellationToken): Promise<vscode.FoldingRange[]> {
-        const id: number = ++abortRequestId;
-        const params: GetFoldingRangesParams = {
-            id: id,
-            uri: document.uri.toString()
-        };
-        return new Promise<vscode.FoldingRange[]>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                this.client.languageClient.sendRequest(GetFoldingRangesRequest, params)
-                    .then((ranges) => {
-                        if (ranges.canceled) {
-                            reject();
-                        } else {
-                            const result: vscode.FoldingRange[] = [];
-                            ranges.ranges.forEach((r) => {
-                                const foldingRange: vscode.FoldingRange = {
-                                    start: r.range.start.line,
-                                    end: r.range.end.line
-                                };
-                                switch (r.kind) {
-                                    case FoldingRangeKind.Comment:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Comment;
-                                        break;
-                                    case FoldingRangeKind.Imports:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Imports;
-                                        break;
-                                    case FoldingRangeKind.Region:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Region;
-                                        break;
-                                    default:
-                                        break;
-                                }
-                                result.push(foldingRange);
-                            });
-                            resolve(result);
-                        }
-                    });
-                token.onCancellationRequested(e => this.client.abortRequest(id));
-            });
-        });
-    }
-}
-
-class SemanticTokensProvider implements vscode.DocumentSemanticTokensProvider {
-    private client: DefaultClient;
-    public onDidChangeSemanticTokensEvent = new vscode.EventEmitter<void>();
-    public onDidChangeSemanticTokens?: vscode.Event<void>;
-    private tokenCaches: Map<string, [number, vscode.SemanticTokens]> = new Map<string, [number, vscode.SemanticTokens]>();
-
-    constructor(client: DefaultClient) {
-        this.client = client;
-        this.onDidChangeSemanticTokens = this.onDidChangeSemanticTokensEvent.event;
-    }
-
-    public async provideDocumentSemanticTokens(document: vscode.TextDocument, token: vscode.CancellationToken): Promise<vscode.SemanticTokens> {
-        return new Promise<vscode.SemanticTokens>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                const uriString: string = document.uri.toString();
-                // First check the token cache to see if we already have results for that file and version
-                const cache: [number, vscode.SemanticTokens] | undefined = this.tokenCaches.get(uriString);
-                if (cache && cache[0] === document.version) {
-                    resolve(cache[1]);
-                } else {
-                    const id: number = ++abortRequestId;
-                    const params: GetSemanticTokensParams = {
-                        id: id,
-                        uri: uriString
-                    };
-                    this.client.languageClient.sendRequest(GetSemanticTokensRequest, params)
-                        .then((tokensResult) => {
-                            if (tokensResult.canceled) {
-                                reject();
-                            } else {
-                                if (tokensResult.fileVersion !== openFileVersions.get(uriString)) {
-                                    reject();
-                                } else {
-                                    const builder: vscode.SemanticTokensBuilder = new vscode.SemanticTokensBuilder(this.client.semanticTokensLegend);
-                                    tokensResult.tokens.forEach((token) => {
-                                        builder.push(token.line, token.character, token.length, token.type, token.modifiers);
-                                    });
-                                    const tokens: vscode.SemanticTokens = builder.build();
-                                    this.tokenCaches.set(uriString, [tokensResult.fileVersion, tokens]);
-                                    resolve(tokens);
-                                }
-                            }
-                        });
-                    token.onCancellationRequested(e => this.client.abortRequest(id));
-                }
-            });
-        });
-    }
-
-    public invalidateFile(uri: string): void {
-        this.tokenCaches.delete(uri);
-        this.onDidChangeSemanticTokensEvent.fire();
-    }
-}
-
-
-class DocumentFormattingEditProvider implements vscode.DocumentFormattingEditProvider {
-    private client: DefaultClient;
-    constructor(client: DefaultClient) {
-        this.client = client;
-    }
-
-    public provideDocumentFormattingEdits(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): Promise<vscode.TextEdit[]> {
-        return new Promise<vscode.TextEdit[]>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                const filePath: string = document.uri.fsPath;
-                const configCallBack = (editorConfigSettings: any| undefined) => {
-                    const params: FormatParams = {
-                        settings: { ...editorConfigSettings },
-                        uri: document.uri.toString(),
-                        insertSpaces: options.insertSpaces,
-                        tabSize: options.tabSize,
-                        character: "",
-                        range:  {
-                            start: {
-                                character: 0,
-                                line: 0
-                            },
-                            end: {
-                                character: 0,
-                                line: 0
-                            }
-                        }
-                    };
-                    return this.client.languageClient.sendRequest(DocumentFormatRequest, params)
-                        .then((textEdits) => {
-                            const result: vscode.TextEdit[] = [];
-                            textEdits.forEach((textEdit) => {
-                                result.push({
-                                    range: new vscode.Range(textEdit.range.start.line, textEdit.range.start.character, textEdit.range.end.line, textEdit.range.end.character),
-                                    newText: textEdit.newText
-                                });
-                            });
-                            resolve(result);
-                        });
-                };
-                const settings: CppSettings = new CppSettings();
-                if (settings.formattingEngine !== "vcFormat") {
-                    configCallBack(undefined);
-                } else {
-                    const editorConfigSettings: any = cachedEditorConfigSettings.get(filePath);
-                    if (!editorConfigSettings) {
-                        editorConfig.parse(filePath).then(configCallBack);
-                    } else {
-                        cachedEditorConfigSettings.set(filePath, editorConfigSettings);
-                        configCallBack(editorConfigSettings);
-                    }
-                }
-            });
-        });
-    }
-}
-
-class DocumentRangeFormattingEditProvider implements vscode.DocumentRangeFormattingEditProvider {
-    private client: DefaultClient;
-    constructor(client: DefaultClient) {
-        this.client = client;
-    }
-
-    public provideDocumentRangeFormattingEdits(document: vscode.TextDocument, range: vscode.Range, options: vscode.FormattingOptions, token: vscode.CancellationToken): Promise<vscode.TextEdit[]> {
-        return new Promise<vscode.TextEdit[]>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                const filePath: string = document.uri.fsPath;
-                const configCallBack = (editorConfigSettings: any | undefined) => {
-                    const params: FormatParams = {
-                        settings: { ...editorConfigSettings },
-                        uri: document.uri.toString(),
-                        insertSpaces: options.insertSpaces,
-                        tabSize: options.tabSize,
-                        character: "",
-                        range: {
-                            start: {
-                                character: range.start.character,
-                                line: range.start.line
-                            },
-                            end: {
-                                character: range.end.character,
-                                line: range.end.line
-                            }
-                        }
-                    };
-                    return this.client.languageClient.sendRequest(DocumentFormatRequest, params)
-                        .then((textEdits) => {
-                            const result: vscode.TextEdit[] = [];
-                            textEdits.forEach((textEdit) => {
-                                result.push({
-                                    range: new vscode.Range(textEdit.range.start.line, textEdit.range.start.character, textEdit.range.end.line, textEdit.range.end.character),
-                                    newText: textEdit.newText
-                                });
-                            });
-                            resolve(result);
-                        });
-                };
-                const settings: CppSettings = new CppSettings();
-                if (settings.formattingEngine !== "vcFormat") {
-                    configCallBack(undefined);
-                } else {
-                    const editorConfigSettings: any = cachedEditorConfigSettings.get(filePath);
-                    if (!editorConfigSettings) {
-                        editorConfig.parse(filePath).then(configCallBack);
-                    } else {
-                        cachedEditorConfigSettings.set(filePath, editorConfigSettings);
-                        configCallBack(editorConfigSettings);
-                    }
-                }
-            });
-        });
-    };
-}
-
-class OnTypeFormattingEditProvider implements vscode.OnTypeFormattingEditProvider {
-    private client: DefaultClient;
-    constructor(client: DefaultClient) {
-        this.client = client;
-    }
-
-    public provideOnTypeFormattingEdits(document: vscode.TextDocument, position: vscode.Position, ch: string, options: vscode.FormattingOptions, token: vscode.CancellationToken): Promise<vscode.TextEdit[]> {
-        return new Promise<vscode.TextEdit[]>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                const filePath: string = document.uri.fsPath;
-                const configCallBack = (editorConfigSettings: any | undefined) => {
-                    const params: FormatParams = {
-                        settings: { ...editorConfigSettings },
-                        uri: document.uri.toString(),
-                        insertSpaces: options.insertSpaces,
-                        tabSize: options.tabSize,
-                        character: ch,
-                        range: {
-                            start: {
-                                character: position.character,
-                                line: position.line
-                            },
-                            end: {
-                                character: 0,
-                                line: 0
-                            }
-                        }
-                    };
-                    return this.client.languageClient.sendRequest(DocumentFormatRequest, params)
-                        .then((textEdits) => {
-                            const result: vscode.TextEdit[] = [];
-                            textEdits.forEach((textEdit) => {
-                                result.push({
-                                    range: new vscode.Range(textEdit.range.start.line, textEdit.range.start.character, textEdit.range.end.line, textEdit.range.end.character),
-                                    newText: textEdit.newText
-                                });
-                            });
-                            resolve(result);
-                        });
-                };
-                const settings: CppSettings = new CppSettings();
-                if (settings.formattingEngine !== "vcFormat") {
-                    // If not using vcFormat, only process on-type requests for ';'
-                    if (ch !== ';') {
-                        const result: vscode.TextEdit[] = [];
-                        resolve(result);
-                    } else {
-                        configCallBack(undefined);
-                    }
-                } else {
-                    const editorConfigSettings: any = cachedEditorConfigSettings.get(filePath);
-                    if (!editorConfigSettings) {
-                        editorConfig.parse(filePath).then(configCallBack);
-                    } else {
-                        cachedEditorConfigSettings.set(filePath, editorConfigSettings);
-                        configCallBack(editorConfigSettings);
-                    }
-                }
-            });
-        });
-    }
-}
 
 export class DefaultClient implements Client {
     private innerLanguageClient?: LanguageClient; // The "client" that launches and communicates with our language "server" process.
